--- src/NzbDrone.Core/Notifications/Xbmc/InvalidXbmcVersionException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace NzbDrone.Core.Notifications.Xbmc
-{
-    public class InvalidXbmcVersionException : Exception
-    {
-        public InvalidXbmcVersionException()
-        {
-        }
-
-        public InvalidXbmcVersionException(string message) : base(message)
-        {
-        }
-    }
-}
